requirements.txt
asyncio==3.4.3
telethon==1.24.0
python-dotenv==0.19.2
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
instaloader==4.9.3
